Add a healthcheck to postgres
The relevant step is the test:prepare and spec:prepare tasks. We
"enhance" them (rake-speak) to do the custom wait_for_db task. This
depends on environment so that the database config is loaded.

Note that we only do this "enhancement" if we're in docker, which we
detect through the use of an environment variable which is set in the
compose file.