Fix NoMethodError in set_course
The NoMethodError came from the regex match. If the regex failed to
match the course id, then `.captures` would be called on `nil`. So we
wrap the entire method in a `rescue` block that will result in a 404 if
any error occurs.

Then, to clean things up a bit more, use the `!` version of `find_by` so
that ActiveRecord raises an exception if no record is found. Now, all
error paths will hit the same `rescue`.